made refactoring
