--- P5/Model/Actions.js
@@ -1,35 +1,0 @@
-export const Actions = {
-
-    USER_INIT: 'USER_INIT',
-    USER_DELETE_NOTICE: 'USER_DELETE_NOTICE',
-    USER_LOG_OUT: 'USER_LOG_OUT',
-
-    FORM_REGISTER: 'FORM_REGISTER',
-    FORM_REGISTER_FNAME_UPDATE: 'FORM_REGISTER_FNAME_UPDATE',
-    FORM_REGISTER_LNAME_UPDATE: 'FORM_REGISTER_LNAME_UPDATE',
-    FORM_REGISTER_EMAIL_UPDATE: 'FORM_REGISTER_EMAIL_UPDATE',
-    FORM_REGISTER_PASSWORD_UPDATE: 'FORM_REGISTER_PASSWORD_UPDATE',
-    FORM_REGISTER_CONFIRM_UPDATE: 'FORM_REGISTER_CONFIRM_UPDATE',
-
-    LOG_IN_EMAIL_UPDATE: 'LOG_IN_EMAIL_UPDATE',
-    LOG_IN_PASSWORD_UPDATE: 'LOG_IN_PASSWORD_UPDATE',
-    LOG_IN: 'LOG_IN',
-
-    NOTICE_SET_PROPRIETOR: 'NOTICE_SET_PROPRIETOR',
-    NOTICE_TITLE_UPDATE: 'NOTICE_TITLE_UPDATE',
-    NOTICE_DESCRIPTION_UPDATE: 'NOTICE_DESCRIPTION_UPDATE',
-    NOTICE_METRIC_UPDATE: 'NOTICE_METRIC_UPDATE',
-    NOTICE_ADDRESS_UPDATE: 'NOTICE_ADDRESS_UPDATE',
-    NOTICE_PRICE_UPDATE: 'NOTICE_PRICE_UPDATE',
-    NOTICE_FOR_RENT_CHANGE: 'NOTICE_FOR_RENT_CHANGE',
-    NOTICE_ADD: 'NOTICE_ADD',
-
-    SEARCH_INIT: 'SEARCH_INIT',
-    SEARCH_MIN_PRICE_UPDATE: 'SEARCH_MIN_PRICE_UPDATE',
-    SEARCH_MAX_PRICE_UPDATE: 'SEARCH_MAX_PRICE_UPDATE',
-    SEARCH_MIN_METRIC_UPDATE: 'SEARCH_MIN_METRIC_UPDATE',
-    SEARCH_MAX_METRIC_UPDATE: 'SEARCH_MAX_METRIC_UPDATE',
-    SEARCH_PAGE_CHANGE: 'SEARCH_PAGE_CHANGE',
-    SEARCH_FOR_RENT_CHANGE: 'SEARCH_FOR_RENT_CHANGE',
-    SEARCH : 'SEARCH',
-};
