--- P5/Model/reducers/indexReducer.js
@@ -1,14 +1,0 @@
-import { combineReducers } from 'redux'
-import newUserState from './newUserReducer'
-import userInfoState from './userInfoReducer'
-import logInState from './logInReducer'
-import newNotificationState from './newNotificationReducer'
-import searchState from './searchReducer'
-
-export default combineReducers({
-    newUserState,
-    userInfoState,
-    logInState,
-    newNotificationState,
-    searchState
-})
