+++ P5/src/Cabinet/actions/CabinetActions.js
@@ -1,0 +1,14 @@
+export const CabinetActions = {
+    USER_INIT: 'USER_INIT',
+    USER_DELETE_NOTICE: 'USER_DELETE_NOTICE',
+    USER_LOG_OUT: 'USER_LOG_OUT',
+
+    NOTICE_SET_PROPRIETOR: 'NOTICE_SET_PROPRIETOR',
+    NOTICE_TITLE_UPDATE: 'NOTICE_TITLE_UPDATE',
+    NOTICE_DESCRIPTION_UPDATE: 'NOTICE_DESCRIPTION_UPDATE',
+    NOTICE_METRIC_UPDATE: 'NOTICE_METRIC_UPDATE',
+    NOTICE_ADDRESS_UPDATE: 'NOTICE_ADDRESS_UPDATE',
+    NOTICE_PRICE_UPDATE: 'NOTICE_PRICE_UPDATE',
+    NOTICE_FOR_RENT_CHANGE: 'NOTICE_FOR_RENT_CHANGE',
+    NOTICE_ADD: 'NOTICE_ADD'
+};
