+++ P5/src/LogIn/actions/LogInActions.js
@@ -1,0 +1,6 @@
+export const LogInActions = {
+    LOG_IN_EMAIL_UPDATE: 'LOG_IN_EMAIL_UPDATE',
+    LOG_IN_PASSWORD_UPDATE: 'LOG_IN_PASSWORD_UPDATE',
+    LOG_IN: 'LOG_IN',
+    LOG_IN_INIT: 'LOG_IN_INIT'
+};
