+++ P5/src/Register/actions/RegisterActions.js
@@ -1,0 +1,8 @@
+export const RegisterActions = {
+    FORM_REGISTER: 'FORM_REGISTER',
+    FORM_REGISTER_FNAME_UPDATE: 'FORM_REGISTER_FNAME_UPDATE',
+    FORM_REGISTER_LNAME_UPDATE: 'FORM_REGISTER_LNAME_UPDATE',
+    FORM_REGISTER_EMAIL_UPDATE: 'FORM_REGISTER_EMAIL_UPDATE',
+    FORM_REGISTER_PASSWORD_UPDATE: 'FORM_REGISTER_PASSWORD_UPDATE',
+    FORM_REGISTER_CONFIRM_UPDATE: 'FORM_REGISTER_CONFIRM_UPDATE'
+};
