+++ P5/src/Search/actions/SearchActions.js
@@ -1,0 +1,10 @@
+export const SearchActions = {
+    SEARCH_INIT: 'SEARCH_INIT',
+    SEARCH_MIN_PRICE_UPDATE: 'SEARCH_MIN_PRICE_UPDATE',
+    SEARCH_MAX_PRICE_UPDATE: 'SEARCH_MAX_PRICE_UPDATE',
+    SEARCH_MIN_METRIC_UPDATE: 'SEARCH_MIN_METRIC_UPDATE',
+    SEARCH_MAX_METRIC_UPDATE: 'SEARCH_MAX_METRIC_UPDATE',
+    SEARCH_PAGE_CHANGE: 'SEARCH_PAGE_CHANGE',
+    SEARCH_FOR_RENT_CHANGE: 'SEARCH_FOR_RENT_CHANGE',
+    SEARCH : 'SEARCH'
+};
